fix: clean up migration command code
